feat(5-1): change structure of cities folder
