+++ src/store/cities/actions.js
@@ -1,0 +1,17 @@
+export const ActionType = {
+  SET_CURRENT_CITY: `SET_CURRENT_CITY`,
+  SET_CURRENT_OFFER: `SET_CURRENT_OFFER`,
+};
+
+export const ActionCreator = {
+  setCurrentCity: (city) => ({
+    type: ActionType.SET_CURRENT_CITY,
+    payload: city,
+  }),
+
+  setCurrentOffer: (offer) => ({
+    type: ActionType.SET_CURRENT_OFFER,
+    payload: offer
+  }),
+
+};
