+++ src/store/nearby/selector.js
@@ -2,7 +2,7 @@ import {createSelector} from "reselect";
 
 import {MAX_NEARBY_OFFERS_COUNT} from "../../utils/constants.js";
 import {getCurrentOffers} from "../offers/selector.js";
-import {getCurrentOffer} from "../cities/selector.js";
+import {getCurrentOffer} from "../cities/selectors.js";
 
 export const getNearby = createSelector(
     getCurrentOffers,
