+++ src/store/reducer.js
@@ -1,6 +1,6 @@
 import {combineReducers} from "redux";
 import {reducer as offers} from "./offers/offers.js";
-import {reducer as cities} from "./cities/cities.js";
+import {reducer as cities} from "./cities/reducer.js";
 import {reducer as reviews} from "./reviews/reviews.js";
 import {reducer as nearby} from "./nearby/nearby.js";
 import NameSpace from "./name-space.js";
